+++ src/init/sphere.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   sphere.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: drenassi <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: nsalles <nsalles@student.42perpignan.fr    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/17 13:50:50 by nsalles           #+#    #+#             */
-/*   Updated: 2024/02/17 21:44:45 by drenassi         ###   ########.fr       */
+/*   Updated: 2024/02/18 02:13:34 by nsalles          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ void	destroy_sphere(t_sphere **sphere)
 /*
  *	Create, set the values and return sphere struct.
 */
-static t_sphere	*new_sphere(t_point *pos, float diameter, int color)
+static t_sphere	*new_sphere(t_point *pos, double diameter, int color)
 {
 	t_sphere	*sphere;
 
@@ -75,7 +75,7 @@ static t_sphere	*get_last_sphere(t_sphere *sphere)
 	return (sphere);
 }
 
-void	set_sphere(t_sphere **sphere, t_point *pos, float diam, int clr)
+void	set_sphere(t_sphere **sphere, t_point *pos, double diam, int clr)
 {
 	t_sphere	*new;
 	t_sphere	*tmp;
